Refactoring batching-mechanisms into Batcher
Guided by an objective to  move batching mechanism and queueing request
to generate batches into a diffenrent thread. This commit is in
preparation for this functionality.

First, PCItem from the looks of it is *Batch*. Renamed to reflect the
same. Fingers crossed, hopefully no naming conflicts with marian.
BatchTranslator translates a "Batch" now, instead of
vector<RequestSentence>. Additional data members are setup at Batch to
enable development.

Workflows previously in Service, but more adequate in Batcher are now
moved, preparing to move Batcher/enqueuing of a request into a new
thread making it non-blocking. This will allow service to queue requests
into the batcher thread and exit, without waiting until the full-request
is queued.

Batcher now has a path with and without pcqueue.