Remove marian entities from Response
Towards https://github.com/browsermt/bergamot-translator/issues/77.

Response previously required marian-deep objects like histories and
vocabs for construction. It has been decided that this structure needs
to be exposed and should not have marian-internals. This commit is the
final nail in the coffin of Response as a marian object concept. The
construction responsiblity of Response is now moved to a ResponseBuilder
class.