Constructs to make Response members optional
Towards https://github.com/browsermt/bergamot-translator/issues/77.

Brings in the concept of ResponseOptions, which should be the intended
outcome of TranslationRequest, providing additional options
along with the source string describing how to construct Response.

It's currently typedef-ed to Ptr<marian::Options>, and subject to change
if and when UnifiedAPI can bring it's own proper equivalent.

Through options, ResponseBuilder now has if-guards to construct
QualityScores or Alignments making it a no-op if the options are unset,
bringing in the desired feature, albeit with a reinvented wheel
structure which is not improper.